Fixed a bug in IE8
Given an element which can't have a child according to the spec, e.g.
meta.

And IE8

When the element's appendChild method is invoked.

Then ie throws an exception.

To solve the previous problem, just check the number of the children.

If there isn't any, don't invoke the method.

Thus it becomes the api users' responsibility to not append nodes to
these kinds of elements.